+++ ConfigBundle/src/Form/Type/AccountPasswordType.php
@@ -1,0 +1,59 @@
+<?php
+
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\ConfigBundle\Form\Type;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Security\Core\Validator\Constraints\UserPassword;
+use Symfony\Component\Validator\Constraints\Length;
+use Symfony\Component\Validator\Constraints\NotBlank;
+use Symfony\Component\Validator\Constraints\NotCompromisedPassword;
+use Symfony\Component\Validator\Constraints\PasswordStrength;
+
+// The password change of AccountController: the current one first, so a session left open on a shared computer cannot take the account over, then the new one under the same constraints as the scaffold's reset form
+class AccountPasswordType extends AbstractType
+{
+    // Neither field is mapped, the controller hashing the new password itself
+    #[\Override]
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('currentPassword', PasswordType::class, [
+                'label' => 'label.current_password',
+                'translation_domain' => 'config',
+                'mapped' => false,
+                'attr' => ['autocomplete' => 'current-password'],
+                'constraints' => [
+                    new NotBlank(message: 'text.password_required'),
+                    new UserPassword(message: 'text.current_password_wrong'),
+                ],
+            ])
+            ->add('plainPassword', RepeatedType::class, [
+                'type' => PasswordType::class,
+                'translation_domain' => 'config',
+                'mapped' => false,
+                'options' => ['attr' => ['autocomplete' => 'new-password']],
+                'first_options' => [
+                    'label' => 'label.new_password',
+                    'constraints' => [
+                        new NotBlank(message: 'text.password_required'),
+                        new Length(min: 8, max: 25, minMessage: 'text.password_min_length', maxMessage: 'text.password_max_length'),
+                        new PasswordStrength(),
+                        new NotCompromisedPassword(),
+                    ],
+                ],
+                'second_options' => ['label' => 'label.password_confirm'],
+                'invalid_message' => 'text.password_mismatch',
+            ]);
+    }
+}